requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
Flask-SQLAlchemy
Flask-MySQLdb
Flask-Migrate
Flask-Login
psycopg2
Flask-Script
WTForms==2.1
Jinja2==2.9.4
MarkupSafe==0.23
Werkzeug==0.11.15
click==6.7
itsdangerous==0.24
wsgiref==0.1.2
WTForms-SQLAlchemy
# ----------------------
# Production Server
# ----------------------
gunicorn